Reworks session management and DB auth

+++ components/AppLayout.tsx
@@ -48,13 +48,13 @@ export default function AppLayout(props: Props) {
   const { dbDeck, dbNotes } = data || {};
   const [isPageLoaded, setIsPageLoaded] = useState(false);
   const [deck, setDeck] = useState<DecryptedDeck>();
+  const [deckKey, setDeckKey] = useState('');
   const [processingAccess, setProcessingAccess] = useState(false);
 
   const isOffline = useStore(state => state.isOffline);
-  const deckKey = useStore(state => state.deckKey);
   const darkMode = useStore(state => state.darkMode);
   const shareModalOpen = useStore(state => state.shareModalOpen);
-  const setDeckKey = useStore(state => state.setDeckKey);
+  const setStoreDeckKey = useStore(state => state.setDeckKey);
   const setNotes = useStore(state => state.setNotes);
   const setNoteTree = useStore(state => state.setNoteTree);
   const setUserId = useStore(state => state.setUserId);
@@ -92,7 +92,7 @@ export default function AppLayout(props: Props) {
   const resetDeck = useCallback(
     async (deleteContributor = false) => {
       toast.error('Unable to verify access');
-      if (deleteContributor) {
+      if (deleteContributor && deckId && user?.id) {
         await supabase
           .from<Contributor>('contributors')
           .delete()
@@ -102,8 +102,7 @@ export default function AppLayout(props: Props) {
       await fetch('/api/reset-recent-deck', { method: 'POST' });
       router.push('/');
     },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [],
+    [deckId, user?.id, router],
   );
 
   const decryptDeck = useCallback(
@@ -111,9 +110,10 @@ export default function AppLayout(props: Props) {
       try {
         const { encrypted_string, encrypted_symmetric_key, access_control_conditions } = dbDeck.access_params;
         const deckKey = await decryptWithLit(encrypted_string, encrypted_symmetric_key, access_control_conditions);
+        setDeckKey(deckKey);
         return deckKey;
       } catch (error) {
-        resetDeck(dbDeck.user_id !== user?.id);
+        await resetDeck(dbDeck.user_id !== user?.id);
       }
     },
     [user?.id, resetDeck],
@@ -141,7 +141,6 @@ export default function AppLayout(props: Props) {
     setDeck(decryptedDeck);
     setCollaborativeDeck(decryptedDeck.access_control_conditions.length > 1);
     setAuthorOnlyNotes(decryptedDeck.author_only_notes ?? false);
-    setDeckKey(key);
 
     if (!dbNotes?.length) {
       setIsPageLoaded(true);
@@ -170,6 +169,7 @@ export default function AppLayout(props: Props) {
       return acc;
     }, {});
     setNotes(notesAsObj);
+    setStoreDeckKey(key);
 
     // Set note tree
     if (decryptedDeck.note_tree) {
@@ -202,7 +202,7 @@ export default function AppLayout(props: Props) {
     setNotes,
     setNoteTree,
     setDeckId,
-    setDeckKey,
+    setStoreDeckKey,
     setUserId,
     decryptDeck,
     setAuthorOnlyNotes,
@@ -216,12 +216,12 @@ export default function AppLayout(props: Props) {
     } else if (litError) {
       console.error('Could not connect to Lit network');
       return;
-    } else if (isLoaded && user && litReady && dbDeck && !isPageLoaded) {
+    } else if (isLoaded && user && litReady && dbDeck?.id === deckId && !isPageLoaded) {
       // Initialize data if there is a user and the data has not been initialized yet
       prepareData();
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [router, user, isLoaded, litReady, litError, dbDeck, isPageLoaded]);
+  }, [router, user, isLoaded, litReady, litError, dbDeck, deckId, isPageLoaded]);
 
   const dbNoteTree = useMemo(() => JSON.stringify(dbDeck?.note_tree), [dbDeck?.note_tree]);
